+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### New features
 
-- `\Wizaplace\SDK\Catalog\DeclinationId` now implements `\JsonSerializable`
+ - `\Wizaplace\SDK\Catalog\DeclinationId` now implements `\JsonSerializable`
 
 ### Bugfixes